+++ Models/DTOs/TodayDashboardDtos.cs
@@ -15,7 +15,7 @@ public class TodayHabitCardDto
     public int TypeId { get; set; }
     public int FrequencyType { get; set; }
     public int TargetCount { get; set; }
-    public int CurrentProgress { get; set; }
+    public decimal CurrentProgress { get; set; }
     public bool IsCompletedToday { get; set; }
     public bool IsPeriodMet { get; set; }
     public decimal PointsPerUnit { get; set; }
